Adicionando a função do botão carregar mais
